Update 'create_responsibility_cadastre_wizard' wizard
